readme - update1
"Associate yourself with men of good quality, if you esteem your reputation. Be not apt to relate news, if you know not the truth thereof. Speak no evil of the absent, for it is unjust. Undertake not what you cannot perform, but be careful to keep your promise. There is but one straight course, and that is to seek truth, and pursue it steadily. Nothing but harmony, honesty, industry and frugality are necessary to make us a great and happy nation." George Washington